Package progress 0.1.1

Progress

A library for displaying progress bars, including support for rendering multiple bars at once. Start by describing of a set of progress bars, then begin rendering them to get access to their respective reporting functions.

Progress_unix

This module defines Unix-specific extensions to the Progress library.

Package info

authors

Craig Ferguson <me@craigfe.io>

changes-files

CHANGES.md

depends

alcotest

astring

dune

fmt

mtime

ocaml

odoc

terminal_size

homepage

https://github.com/CraigFe/progress

issues

https://github.com/CraigFe/progress/issues

license-files

LICENSE

maintainers

Craig Ferguson <me@craigfe.io>

online-doc

https://CraigFe.github.io/progress/

readme-files

README.md

repo

git+https://github.com/CraigFe/progress.git

version

0.1.1